pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trace-ca"
version = "0.1.0"
description = "Government of Canada open data platform — ingest, extract, normalize, query"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.25",
    "google-cloud-storage>=2.18",
    "requests>=2.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
extract = [
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
]
normalize = [
    "rapidfuzz>=3.9",
]
agent = [
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "anthropic>=0.39",
    "sqlglot>=25.0",
]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.12",
]

[dependency-groups]
dev = [
    "trace-ca[extract,normalize,agent,dev]",
]

[tool.setuptools.packages.find]
include = ["shared*", "services*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
